Move starlette tests to starlette not fastapi
