+++ tests/framework_fastapi/test_application.py
@@ -12,10 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
 import pytest
-from testing_support.fixtures import dt_enabled, validate_transaction_metrics
-from testing_support.validators.validate_span_events import validate_span_events
+from testing_support.fixtures import validate_transaction_metrics
 
 
 @pytest.mark.parametrize("endpoint,transaction_name", (
@@ -28,52 +26,5 @@ def test_application(app, endpoint, transaction_name):
     def _test():
         response = app.get(endpoint)
         assert response.status == 200
-
-    _test()
-
-
-@dt_enabled
-@pytest.mark.parametrize("endpoint", ("/graphql", "/graphql_sync"))
-def test_graphql_endpoint(app, endpoint):
-    from graphql import __version__ as version
-
-    FRAMEWORK_METRICS = [
-        ("Python/Framework/GraphQL/%s" % version, 1),
-    ]
-    _test_scoped_metrics = [
-        ("GraphQL/resolve/GraphQL/hello", 1),
-        ("GraphQL/operation/GraphQL/query/<anonymous>/hello", 1),
-    ]
-    _test_unscoped_metrics = [
-        ("GraphQL/all", 1),
-        ("GraphQL/GraphQL/all", 1),
-        ("GraphQL/allWeb", 1),
-        ("GraphQL/GraphQL/allWeb", 1),
-    ] + _test_scoped_metrics
-
-    _expected_query_operation_attributes = {
-        "graphql.operation.type": "query",
-        "graphql.operation.name": "<anonymous>",
-        "graphql.operation.query": "{ hello }",
-    }
-    _expected_query_resolver_attributes = {
-        "graphql.field.name": "hello",
-        "graphql.field.parentType": "Query",
-        "graphql.field.path": "hello",
-        "graphql.field.returnType": "String",
-    }
-
-    @validate_span_events(exact_agents=_expected_query_operation_attributes)
-    @validate_span_events(exact_agents=_expected_query_resolver_attributes)
-    @validate_transaction_metrics(
-        "query/<anonymous>/hello",
-        "GraphQL",
-        scoped_metrics=_test_scoped_metrics,
-        rollup_metrics=_test_unscoped_metrics + FRAMEWORK_METRICS,
-    )
-    def _test():
-        response = app.make_request("POST", endpoint, body=json.dumps({"query": "{ hello }"}), headers={"Content-Type": "application/json"})
-        assert response.status == 200
-        assert "Hello!" in response.body.decode("utf-8")
 
     _test()
